chore: add CI
Currently the CI is not using and GPU instances. In case the tests need
them in the future it's just a matter of following the comments in the
CI configuration. You would comment the GPU executor in and make it the
executor for those tests. You'd also want to run the `install-gpu-deps`
target.